+++ packages/cardano-services/src/Program/index.ts
@@ -1,7 +1,5 @@
 export * from './errors';
-export * from './loadHttpServer';
-export * from './loadTxWorker';
-export * from './programs';
+export * from './programs/';
 export * from './ServiceNames';
 export * from './utils';
 export * from './services';
