+++ packages/cardano-services/src/Program/programs/index.ts
@@ -1,6 +1,10 @@
+export * from './httpServer';
+export * from './txWorker';
+
 /**
  * cardano-services programs
  */
+
 export enum Programs {
   HttpServer = 'HTTP server',
   RabbitmqWorker = 'RabbitMQ worker'
